Added offline support

+++ frontend/src/context/notes/NoteState.js
@@ -1,6 +1,6 @@
 import NoteContext from "./noteContext";
 import {useState} from "react";
-import { toast } from 'react-toastify';
+import {toast} from 'react-toastify';
 
 const NoteState = (props) => {
 
@@ -29,6 +29,12 @@ const NoteState = (props) => {
     //Get all notes
     const getNotes = async () => {
         setProgress(75)
+        if (!navigator.onLine) {
+            const n = JSON.parse(localStorage.getItem('notes'));
+            setNotes(n);
+            setProgress(100)
+            return;
+        }
         const response = await fetch(`${process.env.REACT_APP_HOST}/api/note/fetchNotes`, {
             method: "GET",
             headers: {
@@ -39,12 +45,18 @@ const NoteState = (props) => {
 
         const json = await response.json();
         setNotes(json)
+        localStorage.setItem('notes', JSON.stringify(json))
         setProgress(100)
     }
 
     //Add note
     const addNote = async (title, description, tag, color) => {
         setProgress(30)
+        if (!navigator.onLine) {
+            tst("You're offline", "error")
+            setProgress(100)
+            return;
+        }
         const response = await fetch(`${process.env.REACT_APP_HOST}/api/note/addNote`, {
             method: "POST",
             headers: {
@@ -57,20 +69,24 @@ const NoteState = (props) => {
         setProgress(50)
 
         const res = await response.json()
-        tst(res.message,res.type)
-        if(res.type==="success") {
-            const note = res.savedNote
-            setNotes(notes.concat(note))
+        tst(res.message, res.type)
+        if (res.type === "success") {
+            const note = notes.concat(res.savedNote)
+            setNotes(note)
+            localStorage.setItem('notes', JSON.stringify(note))
         }
         setProgress(100)
     }
 
     //Delete note
     const deleteNote = async (id) => {
+        if (!navigator.onLine) {
+            tst("You're offline", "error")
+            return;
+        }
         const newNote = notes.filter((notes) => {
             return notes._id !== id
         })
-        setNotes(newNote)
 
         const response = await fetch(`${process.env.REACT_APP_HOST}/api/note/deleteNote/${id}`, {
             method: "DELETE",
@@ -82,6 +98,10 @@ const NoteState = (props) => {
 
         const res = await response.json();
         tst(res.message, res.type)
+        if(res.type==="success"){
+            setNotes(newNote)
+            localStorage.setItem('notes',JSON.stringify(newNote))
+        }
 
         /*map, filter and reduce , spread operator, multiple function arguments operator.*/
     }
@@ -89,7 +109,10 @@ const NoteState = (props) => {
 
     //Edit note
     const editNote = async (id, title, description, tag, color) => {
-
+        if (!navigator.onLine) {
+            tst("You're offline", "error")
+            return;
+        }
         const newNotes = JSON.parse(JSON.stringify(notes))                          // error when directly changing notes as useState does not allow therefore creating a copy
 
         const response = await fetch(`${process.env.REACT_APP_HOST}/api/note/updateNote/${id}`, {
@@ -101,11 +124,11 @@ const NoteState = (props) => {
             body: JSON.stringify({title, description, tag, color})
         });
         const res = await response.json()
-        if(res.type==="same"){
+        if (res.type === "same") {
             return false;
         }
         tst(res.message, res.type)
-        if(res.type==="success"){
+        if (res.type === "success") {
             for (let i = 0; i < newNotes.length; i++) {
                 const element = newNotes[i]
                 if (element._id === id) {
@@ -117,6 +140,7 @@ const NoteState = (props) => {
                 }
             }
             setNotes(newNotes)
+            localStorage.setItem('notes',JSON.stringify(newNotes))
             return true;
         }
         return false;
@@ -125,7 +149,19 @@ const NoteState = (props) => {
     const [showColors, setShowColors] = useState(false);
 
     return (
-        <NoteContext.Provider value={{notes, setNotes, addNote, deleteNote, editNote, getNotes, showColors, setShowColors, tst, progress, setProgress}}>
+        <NoteContext.Provider value={{
+            notes,
+            setNotes,
+            addNote,
+            deleteNote,
+            editNote,
+            getNotes,
+            showColors,
+            setShowColors,
+            tst,
+            progress,
+            setProgress
+        }}>
             {props.children}
         </NoteContext.Provider>
     )
